Create escrowBank.sol
Escrow is the outsider which holds the asset(asset can be cash, bond, stocks) on the presence of two gatherings. Escrow will deliver the asset when certain circumstances are met.
For Instance, "A" will be a vendor and needs to sell his vehicle, "B" is a purchaser who needs to purchase "A's" vehicle so they will contact Escrow "C"(an judge) which hold the resource until "B" gets the vehicle. At the point when this condition will be met, Escrow will deliver the asset to "A". This tackles the issue of trust and forestalls any inconsistency.
Lets the compose a savvy contract for the Escrow utilizing strength language.
Here Savvy agreement will hold the resource, which will be delivered on when conditions are satisfied.